ats-resume-bot/requirements.txt
python-telegram-bot==20.8
python-dotenv
fpdf
python-docx
PyPDF2
